fix: edit spacing between components in Home, migrate to <PreviousMeetup/>
